Move option parsing from CLI to a new class Options.
