+++ .rubocop.yml
@@ -7,4 +7,4 @@ MethodLength:
   Max: 30
 
 ClassLength:
-  Max: 250
+  Max: 210
