+++ lib/rubocop.rb
@@ -165,4 +165,5 @@ require 'rubocop/token'
 require 'rubocop/processed_source'
 require 'rubocop/source_parser'
 require 'rubocop/cli'
+require 'rubocop/options'
 require 'rubocop/version'
